requirements.txt
# requirements.txt

# 核心依赖
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
pymupdf>=1.22.5
streamlit>=1.29.0
python-dotenv>=1.0.0

# 嵌入和向量存储
sentence-transformers>=2.2.2
chromadb>=0.4.18

# 文档处理
unstructured>=0.10.30
pdf2image>=1.16.3
pytesseract>=0.3.10

# 网页处理
beautifulsoup4>=4.12.2
requests